refactor(tools): merge makefolder into tools
Resolves: #12
