--- utils/makefolder.py
@@ -1,6 +1,0 @@
-import os
-
-def makefolder(*args):
-    dir = os.path.join(*args)
-    if not os.path.exists(dir):
-        os.makedirs(dir)
